+++ docs/changelog.md
@@ -11,6 +11,8 @@
 ###Fixed
 
   - Encoding URL's
+  - Client
+    - Only debug log level
 
 ###Updated